structured the services and fixed layout
