+++ frontend/src/components/Layout.jsx
@@ -1,0 +1,12 @@
+import NavBar from "./NavBar";
+
+function Layout ({children}) {
+    return(
+       <div>
+        <NavBar />
+        <div>{children}</div>
+       </div>
+    )
+}
+
+export default Layout;
